Revert "ignore stale qemu.pid files" and give isotovideo more time
to kill qemu - we'll see

+++ lib/OpenQA/Worker/Pool.pm
@@ -43,6 +43,10 @@ sub lockit(){
 
 sub clean_pool(){
     return if $nocleanup;
+    if (-e "$pooldir/qemu.pid") {
+        print "QEMU should be dead - WASUP?\n";
+        exit(1);
+    }
     for my $file (<$pooldir/*>) {
         if (-d $file) {
             remove_tree($file);
